Refactor: rename qkv_proj_rope.py to decode_qkv_proj_rope.py
Disambiguates the decode-side kernel from the new prefill_qkv_proj_rope
module. Updates all `from qkv_proj_rope import ...` callers (3 decode
attention files, 1 prefill draft, and prefill_qkv_proj_rope itself).
Function names (qkv_proj_rope, golden_qkv_proj_rope) are unchanged.
